Require an agreement before showing the application form
